multiple drafts support
