+++ lib/draftsman/single/draft.rb
@@ -1,0 +1,7 @@
+require 'draftsman/shared_draft_methods.rb'
+
+class Draftsman::Single::Draft < Draftsman::SharedDraftMethods
+
+  belongs_to :item, polymorphic: true
+
+end
